Changed DataLoad to have Double.type for label. Realised that the features vector is just the features for a single point and transform() does the rest. We have a result!

+++ Spark_MLPipeline/src/main/scala-2.11/dev/data_load/DataLoad.scala
@@ -33,7 +33,7 @@ object DataLoad {
         attributes(14).toDouble, attributes(15).toDouble, attributes(16).toDouble,
         attributes(17).toDouble, attributes(18).toDouble, attributes(19).toDouble,
         attributes(20).toDouble, attributes(21).toDouble, attributes(22).toDouble,
-        attributes(23).toInt))
+        attributes(23).toDouble))
       .toDF()
       .cache()).toOption
     df
